src/backend/pyproject.toml
[project]
name = "find-your-tribe-backend"
version = "0.1.0"
description = "Backend GraphQL API for find-your-tribe"
requires-python = ">=3.12"

dependencies = [
    "fastapi[standard]",
    "strawberry-graphql[fastapi]",
    "sqlalchemy[asyncio]>=2.0",
    "asyncpg",
    "alembic",
    "pgvector",
    "python-ulid",
    "pydantic-settings",
    "uvicorn[standard]",
    "bcrypt",
    "PyJWT",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "httpx",
    "ruff",
]

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4", # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501", # line too long (handled by formatter)
]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]
